Put methods where they belong. First test
