+++ collector/mysql_connect.py
@@ -9,7 +9,7 @@ class Mysql_Connect(object):
 		if Mysql_Connect.conn is None:
 			Mysql_Connect.conn = MySQLdb.connect(
 				host = '127.0.0.1',
-				port = 3306,
+				port = 3307,
 				user = 'adminw9XD8Ju',
 				passwd = 'QIY_8QyGMW-X', 
 				db = 'ucd')
