+++ main_parse.py
@@ -1,0 +1,7 @@
+import processor.cloud_parser as pars
+import generator.cloud_generator as gen
+
+
+connection = pars.Cloud_Parser()
+generator = gen.Cloud_Generator(64,64)
+connection.get_data()
